feat(app): update default monospace font to FiraCode
Updated the default monospace font across the application from 'IBM Plex Mono' to 'FiraCode'. This change affects both the main Nuxt configuration and the UnoCSS configuration to ensure consistency in typography across the platform. This update aims to enhance the readability and aesthetics of code elements and UI components that utilize the monospace font.