chore(docker): add nano-core and nano-module-core services
Added new `nano-core` and `nano-module-core` service configurations to the `docker-compose.yaml`. These services are set up with the `cimg/node:20.17.0` image and configured to run development commands for the respective `@nwrx/core` and `@nwrx/module-core` packages. This will facilitate local development and testing by ensuring the services are properly isolated and configured in the Docker environment.