chore(monorepo): update dependencies and configurations across packages
Updated various package dependencies and configurations to streamline development and ensure compatibility. Major changes include the addition of new dependencies such as `@microsoft/api-extractor` and `@swc/core` to the root `package.json`, and integration of `@nwrx/module-google` into the `api` package. Adjustments were also made to the `core` and `module-core` packages, including reorganization of dependencies and update of build scripts from `rollup` to `tsup` for enhanced build processes.

These updates are part of ongoing efforts to maintain and improve the build system and dependency management in our monorepo setup.