chore(app): added module-piste-gouv-fr import to server initialization
Imported the `module-piste-gouv-fr` in the server index to ensure it's included during the application initialization. This allows us to leverage the functionalities of the Legifrance integration from the newly added module. The ESLint rule for sorting imports has also been disabled to maintain the desired import order without linting interruptions.